+++ src/main/java/Task.java
@@ -1,3 +1,7 @@
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
 public class Task {
     protected String description;
     protected String subDescription; // For /by, /at
@@ -47,8 +51,21 @@ class ToDo extends Task {
 }
 
 class Deadline extends Task {
-    public Deadline(String description, String by) {
+    public Deadline(String description, String by) throws DukeException {
         super(description, by);
+        // If deadline /by is of correct format
+        if (isValidDateTimeFormat(by)) {
+            try {
+                SimpleDateFormat displayFormat = new SimpleDateFormat("dd MMMM yyyy, hh:mm a");
+                SimpleDateFormat inputFormat = new SimpleDateFormat("dd/MM/yyyy HHmm");
+                Date date = inputFormat.parse(by);
+                // Output: 2nd of December 2019, 6pm
+                by = displayFormat.format(date);
+                this.subDescription = by;
+            } catch (ParseException e) {
+                throw new DukeException("Unable to recognise date-time provided.");
+            }
+        }
     }
 
     @Override
@@ -56,6 +73,22 @@ class Deadline extends Task {
         return "[D]" + super.toString() + " (by: " + subDescription + ")";
     }
     public String getTaskType() { return "D"; }
+
+    // Returns true if string is of format 2/12/2019 1800
+    private boolean isValidDateTimeFormat(String str) {
+        SimpleDateFormat inputFormat = new SimpleDateFormat("dd/MM/yyy hhmm");
+        try {
+            /*str.matches("([1-9]|[1-2][0-9]|3[0-1])/" +
+                    "([1-9]|1[0-2])/" +
+                    "([0-9]{4})" +
+                    "\\s[0-2][0-9]([0-5]{2})")
+             */
+            inputFormat.parse(str);
+            return true;
+        } catch (ParseException e) {
+            return false;
+        }
+    }
 
 }
